pyproject.toml
[project]
name = "hackersky"
version = "0.1.0"
description = "A Bluesky bot that posts trending news.ycombinator.com articles"
authors = [
    { name = "Aaron Mamparo", email = "aaronmamparo@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.13,<3.14"
package-mode = false
dependencies = ["python-dotenv (>=1.1.0,<2.0.0)", "boto3 (>=1.37.37,<2.0.0)", "injector (>=0.22.0,<0.23.0)", "requests (>=2.32.3,<3.0.0)", "beautifulsoup4 (>=4.13.4,<5.0.0)", "fake-useragent (>=2.2.0,<3.0.0)", "pillow (>=11.2.1,<12.0.0)", "tqdm (>=4.67.1,<5.0.0)", "rss-parser (>=2.1.0,<3.0.0)", "atproto (>=0.0.61,<0.0.62)"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pylint = "^3.3.6"
mypy = "^1.15.0"
aws-cdk-lib = "^2.190.0"

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"

